requirements.txt
certifi==2020.12.5
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.15
django-allauth==0.42.0
django-ckeditor==6.0.0
django-filter==2.4.0
django-js-asset==1.2.2
django-mdeditor==0.1.18
django-model-utils==4.0.0
django-mptt==0.11.0
django-notifications-hq==1.6.0
django-password-reset==2.0
django-redis==4.12.1
django-shortuuidfield==0.1.3
django-simpleui==2020.9
django-taggit==1.3.0
djangorestframework==3.12.2
drf-extensions==0.7.0
idna==2.10
importlib-metadata==2.0.0
itsdangerous==1.1.0
jsonfield==3.1.0
Markdown==3.2.2
MarkupSafe==1.1.1
oauthlib==3.1.0
Pillow==7.2.0
Pygments==2.7.1
PyMySQL==0.10.1
PyPDF2==1.26.0
PyQt5==5.15.2
pyqt5-plugins==5.15.2.2.0.1
PyQt5-Qt==5.15.2
PyQt5-sip==12.8.1
pyqt5-tools==5.15.2.3.0.2
python-dateutil==2.8.1
python-dotenv==0.15.0
python-memcached==1.59
python3-openid==3.2.0
pytz==2020.1
pywin32==300
qiniu==7.3.1
qt5-applications==5.15.2.2.1
qt5-tools==5.15.2.1.0.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
shortuuid==1.0.1
six==1.15.0
sqlparse==0.3.1
swapper==1.1.2.post1
TimeConvert==1.5.2
tzlocal==2.1
urllib3==1.25.10
wincertstore==0.2
zipp==3.2.0
